main.py
import os 
from dotenv import load_dotenv
from google import genai
from argparse import ArgumentParser
from google.genai import types

load_dotenv()
api_key = os.environ.get('GEMINI_API_KEY')

if api_key is None:
    raise RuntimeError("no required env variable")

client = genai.Client(api_key=api_key)

content: str = "Why is Boot.dev such a great place to learn backend development? Use one paragraph maximum."

def main():
    print("Hello from bootdev-ai-agent!")

    parser = ArgumentParser(description="chatbot")
    parser.add_argument("user_prompt", type=str, help="User prompt")
    parser.add_argument("--verbose", action="store_true", help="Enable verbose output")
    args = parser.parse_args()

    messages = [types.Content(role="user", parts=[types.Part(text=args.user_prompt)])]

    response = client.models.generate_content(model='gemini-2.5-flash', contents=messages)

    if args.verbose:
        print(f"User prompt: {args.user_prompt}")
        if response.usage_metadata is not None:
            print(f"Prompt tokens: {response.usage_metadata.prompt_token_count}")
            print(f"Response tokens: {response.usage_metadata.candidates_token_count}")
        else:
            raise RuntimeError("failed API request")    

    print("Response:")
    print(response.text)


if __name__ == "__main__":
    main()
