pyproject.toml
[project]
name = "bootdev-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai==1.12.1",
    "pytest>=9.0.3",
    "python-dotenv==1.1.0",
]
